Fix for "conn busy" error.

+++ internal/storage/withdrawals.go
@@ -10,7 +10,7 @@ var ErrNotEnoughBalance = errors.New("not enough balance")
 
 type BalanceInfo struct {
 	Current   float64 `json:"current"`
-	Withdrawn int64   `json:"withdrawn"`
+	Withdrawn float64 `json:"withdrawn"`
 }
 
 type Withdrawal struct {
